+++ src/net/tmwa/partyhandler.cpp
@@ -193,8 +193,8 @@ void PartyHandler::setShareExperience(PartyShare share)
         return;
 
     MessageOut outMsg(CMSG_PARTY_SETTINGS);
-    outMsg.writeInt16(share);
-    outMsg.writeInt16(mShareItems);
+    outMsg.writeInt16(static_cast<int16_t>(share));
+    outMsg.writeInt16(static_cast<int16_t>(mShareItems));
 }
 
 void PartyHandler::setShareItems(PartyShare share)
@@ -203,8 +203,8 @@ void PartyHandler::setShareItems(PartyShare share)
         return;
 
     MessageOut outMsg(CMSG_PARTY_SETTINGS);
-    outMsg.writeInt16(mShareExp);
-    outMsg.writeInt16(share);
+    outMsg.writeInt16(static_cast<int16_t>(mShareExp));
+    outMsg.writeInt16(static_cast<int16_t>(share));
 }
 
 } // namespace TmwAthena
